requirements.txt
# Flask 核心依赖
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1

# 生产环境 WSGI 服务器
gunicorn==21.2.0

# 数据验证
email_validator==2.1.0.post1

# Markdown 和代码高亮
markdown==3.5.1
Pygments==2.19.2

# 安全防护
bleach==6.1.0

# 缓存
Flask-Caching==2.1.0

# 评论系统（前端CDN引入，此处仅记录）

# RSS 生成
feedgen==0.9.0

# 图片处理
Pillow>=10.1.0

# PostgreSQL 数据库（Render 部署使用）
psycopg2-binary==2.9.10
